+++ src/Command/DeleteAllCommand.php
@@ -23,6 +23,7 @@ use DatabaseBackup\Utility\BackupManager;
 
 /**
  * Deletes all backup files
+ * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#delete_all
  */
 class DeleteAllCommand extends Command
 {
@@ -42,7 +43,6 @@ class DeleteAllCommand extends Command
      * @param \Cake\Console\ConsoleIo $io The console io
      * @return void
      * @throws \ErrorException|\ReflectionException
-     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#delete_all
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {