+++ src/Command/RotateCommand.php
@@ -24,6 +24,7 @@ use Exception;
 
 /**
  * Rotates backups
+ * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#rotate
  */
 class RotateCommand extends Command
 {
@@ -49,7 +50,6 @@ class RotateCommand extends Command
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
      * @return void
-     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#rotate
      * @throws \Cake\Console\Exception\StopException|\ReflectionException
      */
     public function execute(Arguments $args, ConsoleIo $io): void