+++ tests/TestCase/Driver/SqliteTest.php
@@ -1,5 +1,4 @@
 <?php
-/** @noinspection PhpDocMissingThrowsInspection, PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -15,7 +14,6 @@ declare(strict_types=1);
  */
 namespace DatabaseBackup\Test\TestCase\Driver;
 
-use DatabaseBackup\Driver\Sqlite;
 use DatabaseBackup\TestSuite\DriverTestCase;
 
 /**
@@ -31,8 +29,8 @@ class SqliteTest extends DriverTestCase
     {
         parent::setUp();
 
-        if (!$this->Driver instanceof Sqlite) {
-            $this->markTestSkipped('Skipping tests for Sqlite, current driver is ' . $this->Driver->getDriverName());
+        if ($this->getConnection()->config()['scheme'] !== 'sqlite') {
+            $this->markTestSkipped('Skipping tests for sqlite, current driver is `' . $this->getConnection()->config()['scheme'] . '`');
         }
     }
 }