+++ tests/TestCase/Utility/BackupExportTest.php
@@ -16,10 +16,13 @@ declare(strict_types=1);
 namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
+use Cake\Event\EventList;
 use Cake\TestSuite\EmailTrait;
+use DatabaseBackup\Driver\Sqlite;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
-use Tools\TestSuite\ReflectionTrait;
+use Symfony\Component\Process\Exception\ProcessTimedOutException;
+use Symfony\Component\Process\Process;
 
 /**
  * BackupExportTest class
@@ -27,7 +30,6 @@ use Tools\TestSuite\ReflectionTrait;
 class BackupExportTest extends TestCase
 {
     use EmailTrait;
-    use ReflectionTrait;
 
     /**
      * @var \DatabaseBackup\Utility\BackupExport
@@ -43,6 +45,7 @@ class BackupExportTest extends TestCase
         parent::setUp();
 
         $this->BackupExport ??= new BackupExport();
+        $this->BackupExport->getDriver()->getEventManager()->setEventList(new EventList());
     }
 
     /**
@@ -53,8 +56,8 @@ class BackupExportTest extends TestCase
     public function testCompression(): void
     {
         $this->BackupExport->compression('bzip2');
-        $this->assertSame('bzip2', $this->getProperty($this->BackupExport, 'compression'));
-        $this->assertSame('sql.bz2', $this->getProperty($this->BackupExport, 'extension'));
+        $this->assertSame('bzip2', $this->BackupExport->compression);
+        $this->assertSame('sql.bz2', $this->BackupExport->extension);
 
         //With an invalid type
         $this->expectExceptionMessage('Invalid compression type');
@@ -69,31 +72,31 @@ class BackupExportTest extends TestCase
     public function testFilename(): void
     {
         $this->BackupExport->filename('backup.sql.bz2');
-        $this->assertSame(Configure::read('DatabaseBackup.target') . 'backup.sql.bz2', $this->getProperty($this->BackupExport, 'filename'));
-        $this->assertSame('bzip2', $this->getProperty($this->BackupExport, 'compression'));
-        $this->assertSame('sql.bz2', $this->getProperty($this->BackupExport, 'extension'));
+        $this->assertSame(Configure::read('DatabaseBackup.target') . 'backup.sql.bz2', $this->BackupExport->filename);
+        $this->assertSame('bzip2', $this->BackupExport->compression);
+        $this->assertSame('sql.bz2', $this->BackupExport->extension);
 
         //Compression is ignored, because there's a filename
         $this->BackupExport->compression('gzip')->filename('backup.sql.bz2');
-        $this->assertSame('backup.sql.bz2', basename($this->getProperty($this->BackupExport, 'filename')));
-        $this->assertSame('bzip2', $this->getProperty($this->BackupExport, 'compression'));
-        $this->assertSame('sql.bz2', $this->getProperty($this->BackupExport, 'extension'));
+        $this->assertSame('backup.sql.bz2', basename($this->BackupExport->filename));
+        $this->assertSame('bzip2', $this->BackupExport->compression);
+        $this->assertSame('sql.bz2', $this->BackupExport->extension);
 
         //Filename with `{$DATABASE}` pattern
         $this->BackupExport->filename('{$DATABASE}.sql');
-        $this->assertSame('test.sql', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertSame('test.sql', basename($this->BackupExport->filename));
 
         //Filename with `{$DATETIME}` pattern
         $this->BackupExport->filename('{$DATETIME}.sql');
-        $this->assertMatchesRegularExpression('/^\d{14}\.sql$/', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertMatchesRegularExpression('/^\d{14}\.sql$/', basename($this->BackupExport->filename));
 
         //Filename with `{$HOSTNAME}` pattern
         $this->BackupExport->filename('{$HOSTNAME}.sql');
-        $this->assertSame('localhost.sql', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertSame('localhost.sql', basename($this->BackupExport->filename));
 
         //Filename with `{$TIMESTAMP}` pattern
         $this->BackupExport->filename('{$TIMESTAMP}.sql');
-        $this->assertMatchesRegularExpression('/^\d{10}\.sql$/', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertMatchesRegularExpression('/^\d{10}\.sql$/', basename($this->BackupExport->filename));
 
         //With invalid extension
         $this->expectExceptionMessage('Invalid `txt` file extension');
@@ -107,7 +110,7 @@ class BackupExportTest extends TestCase
     public function testRotate(): void
     {
         $this->BackupExport->rotate(10);
-        $this->assertSame(10, $this->getProperty($this->BackupExport, 'rotate'));
+        $this->assertSame(10, $this->BackupExport->rotate);
 
         $this->expectExceptionMessage('Invalid rotate value');
         $this->BackupExport->rotate(-1)->export();
@@ -120,11 +123,21 @@ class BackupExportTest extends TestCase
     public function testSend(): void
     {
         $this->BackupExport->send();
-        $this->assertNull($this->getProperty($this->BackupExport, 'emailRecipient'));
+        $this->assertNull($this->BackupExport->emailRecipient);
 
         $recipient = 'recipient@example.com';
         $this->BackupExport->send($recipient);
-        $this->assertSame($recipient, $this->getProperty($this->BackupExport, 'emailRecipient'));
+        $this->assertSame($recipient, $this->BackupExport->emailRecipient);
+    }
+
+    /**
+     * @test
+     * @uses \DatabaseBackup\Utility\BackupExport::timeout()
+     */
+    public function testTimeout(): void
+    {
+        $this->BackupExport->timeout(120);
+        $this->assertSame(120, $this->BackupExport->timeout);
     }
 
     /**
@@ -133,23 +146,25 @@ class BackupExportTest extends TestCase
      */
     public function testExport(): void
     {
-        $file = $this->BackupExport->export();
+        $file = $this->BackupExport->export() ?: '';
         $this->assertFileExists($file);
         $this->assertMatchesRegularExpression('/^backup_test_\d{14}\.sql$/', basename($file));
+        $this->assertEventFired('Backup.beforeExport', $this->BackupExport->getDriver()->getEventManager());
+        $this->assertEventFired('Backup.afterExport', $this->BackupExport->getDriver()->getEventManager());
 
         //Exports with `compression()`
-        $file = $this->BackupExport->compression('bzip2')->export();
+        $file = $this->BackupExport->compression('bzip2')->export() ?: '';
         $this->assertFileExists($file);
         $this->assertMatchesRegularExpression('/^backup_test_\d{14}\.sql\.bz2$/', basename($file));
 
         //Exports with `filename()`
-        $file = $this->BackupExport->filename('backup.sql.bz2')->export();
+        $file = $this->BackupExport->filename('backup.sql.bz2')->export() ?: '';
         $this->assertFileExists($file);
         $this->assertSame('backup.sql.bz2', basename($file));
 
         //Exports with `send()`
         $recipient = 'recipient@example.com';
-        $file = $this->BackupExport->filename('exportWithSend.sql')->send($recipient)->export();
+        $file = $this->BackupExport->filename('exportWithSend.sql')->send($recipient)->export() ?: '';
         $this->assertMailSentFrom(Configure::readOrFail('DatabaseBackup.mailSender'));
         $this->assertMailSentTo($recipient);
         $this->assertMailSentWith('Database backup ' . basename($file) . ' from localhost', 'subject');
@@ -168,11 +183,57 @@ class BackupExportTest extends TestCase
      */
     public function testExportWithDifferentChmod(): void
     {
-        $file = $this->BackupExport->filename('exportWithNormalChmod.sql')->export();
+        $file = $this->BackupExport->filename('exportWithNormalChmod.sql')->export() ?: '';
         $this->assertSame('0664', substr(sprintf('%o', fileperms($file)), -4));
 
         Configure::write('DatabaseBackup.chmod', 0777);
-        $file = $this->BackupExport->filename('exportWithDifferentChmod.sql')->export();
+        $file = $this->BackupExport->filename('exportWithDifferentChmod.sql')->export() ?: '';
         $this->assertSame('0777', substr(sprintf('%o', fileperms($file)), -4));
+    }
+
+    /**
+     * Test for `export()` method. Export is stopped by the `Backup.beforeExport` event (implemented by driver)
+     * @test
+     * @uses \DatabaseBackup\Utility\BackupExport::export()
+     */
+    public function testExportStoppedByBeforeExport(): void
+    {
+        $Driver = $this->createPartialMock(Sqlite::class, ['beforeExport']);
+        $Driver->method('beforeExport')->willReturn(false);
+        $Driver->getEventManager()->on($Driver);
+        $BackupExport = $this->createPartialMock(BackupExport::class, ['getDriver']);
+        $BackupExport->method('getDriver')->willReturn($Driver);
+        $this->assertFalse($BackupExport->export());
+    }
+
+    /**
+     * Test for `export()` method, on failure (error for `Process`)
+     * @test
+     * @uses \DatabaseBackup\Utility\BackupExport::export()
+     */
+    public function testExportOnFailure(): void
+    {
+        $expectedError = 'mysqldump: Got error: 1044: "Access denied for user \'root\'@\'localhost\' to database \'noExisting\'" when selecting the database';
+        $this->expectExceptionMessage('Export failed with error message: `' . $expectedError . '`');
+        $Process = $this->createConfiguredMock(Process::class, ['getErrorOutput' => $expectedError . PHP_EOL, 'isSuccessful' => false]);
+        $BackupExport = $this->createPartialMock(BackupExport::class, ['getProcess']);
+        $BackupExport->method('getProcess')->willReturn($Process);
+        $BackupExport->export();
+    }
+
+    /**
+     * Test for `export()` method, exceeding the timeout
+     * @see https://symfony.com/doc/current/components/process.html#process-timeout
+     * @test
+     * @uses \DatabaseBackup\Utility\BackupExport::export()
+     */
+    public function testExportExceedingTimeout(): void
+    {
+        $this->expectException(ProcessTimedOutException::class);
+        $this->expectExceptionMessage('The process "dir" exceeded the timeout of 60 seconds');
+        $ProcessTimedOutException = new ProcessTimedOutException(Process::fromShellCommandline('dir'), 1);
+        $BackupExport = $this->createPartialMock(BackupExport::class, ['getProcess']);
+        $BackupExport->method('getProcess')->willThrowException($ProcessTimedOutException);
+        $BackupExport->export();
     }
 }